userservice/requirements.txt
flask
flask-sqlalchemy
pymysql
flask-bcrypt
flask-jwt-extended
python-dotenv
